fix: Player의 생성자에서 '딜러'이름을 예외처리하도록 수정
- Dealer의 생성자에서 Name객체를 '딜러'라는 이름으로 생성하므로, Name에서 예외처리 시 무조건 에러 발생
