+++ src/main/java/blackjack/domain/user/Player.java
@@ -6,10 +6,19 @@ import blackjack.domain.result.WinningStatus;
 
 public class Player extends User {
 
+
+    static final String NAME_IS_DEALER_NAME_EXCEPTION_MESSAGE = "플레이어의 이름은 '딜러'일 수 없습니다.";
     private static final int FIRST_OPEN_CARD_COUNT = 2;
 
-    protected Player(String name, CardGroup cardGroup) {
+    protected Player(final String name, final CardGroup cardGroup) {
         super(name, cardGroup);
+        validateIsDealerName(name);
+    }
+
+    private void validateIsDealerName(final String name) {
+        if (Dealer.DEALER_NAME.equals(name)) {
+            throw new IllegalArgumentException(NAME_IS_DEALER_NAME_EXCEPTION_MESSAGE);
+        }
     }
 
     @Override
